Made RGBAColor components 8 bits wide
The values are all in the range 0..255, so storing more than 8 bits
is wasteful.

+++ src/surface.h
@@ -25,9 +25,10 @@
 
 #include <SDL.h>
 #include <string>
+#include <cstdint>
 
 struct RGBAColor {
-	unsigned short r,g,b,a;
+	uint8_t r, g, b, a;
 };
 
 RGBAColor strtorgba(const std::string &strColor);
